backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.6"
dependencies = [
    "aio-pika>=9.5.8",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "beautifulsoup4>=4.14.3",
    "cryptography>=46.0.5",
    "fastapi>=0.129.0",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.13.0",
    "pydantic[email]>=2.12.5",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.45",
    "structlog>=25.5.0",
    "uvicorn>=0.41.0",
]
